Move TestRedirectBase to utils.
