+++ tests/test_downloadermiddleware_redirect.py
@@ -16,11 +16,10 @@ from scrapy.spiders import Spider
 from scrapy.utils.misc import build_from_crawler
 from scrapy.utils.spider import DefaultSpider
 from scrapy.utils.test import get_crawler
-from tests.test_downloadermiddleware_redirect_base import Base
-from tests.utils.redirect import REDIRECT_SCHEME_CASES, SCHEME_PARAMS
+from tests.utils.redirect import REDIRECT_SCHEME_CASES, SCHEME_PARAMS, TestRedirectBase
 
 
-class TestRedirectMiddleware(Base.Test):
+class TestRedirectMiddleware(TestRedirectBase):
     mwcls = RedirectMiddleware
     reason = 302
