+++ tests/test_downloadermiddleware_redirect_metarefresh.py
@@ -12,12 +12,12 @@ from scrapy.http import HtmlResponse, Request, Response
 from scrapy.spiders import Spider
 from scrapy.utils.misc import build_from_crawler
 from scrapy.utils.test import get_crawler
-from tests.test_downloadermiddleware_redirect_base import Base
 from tests.utils.redirect import (
     HTTP_SCHEMES,
     NON_HTTP_SCHEMES,
     REDIRECT_SCHEME_CASES,
     SCHEME_PARAMS,
+    TestRedirectBase,
 )
 
 
@@ -26,7 +26,7 @@ def meta_refresh_body(url, interval=5):
     return html.encode("utf-8")
 
 
-class TestMetaRefreshMiddleware(Base.Test):
+class TestMetaRefreshMiddleware(TestRedirectBase):
     mwcls = MetaRefreshMiddleware
     reason = "meta refresh"
